rubocop autofixed 500+ offenses

+++ db/migrate/20200328094816_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This migration comes from acts_as_taggable_on_engine (originally 2)
 class AddMissingUniqueIndices < ActiveRecord::Migration[6.0]
   disable_ddl_transaction!
@@ -5,11 +7,13 @@ class AddMissingUniqueIndices < ActiveRecord::Migration[6.0]
   def self.up
     add_index ActsAsTaggableOn.tags_table, :name, unique: true, algorithm: :concurrently
 
-    remove_index ActsAsTaggableOn.taggings_table,
-                 :tag_id if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
+    if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
+      remove_index ActsAsTaggableOn.taggings_table,
+                   :tag_id
+    end
     remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
     add_index ActsAsTaggableOn.taggings_table,
-              [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
+              %i[tag_id taggable_id taggable_type context tagger_id tagger_type],
               unique: true,
               name: 'taggings_idx',
               algorithm: :concurrently
@@ -20,11 +24,13 @@ class AddMissingUniqueIndices < ActiveRecord::Migration[6.0]
 
     remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_idx'
 
+    unless index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
+      add_index ActsAsTaggableOn.taggings_table,
+                :tag_id,
+                algorithm: :concurrently
+    end
     add_index ActsAsTaggableOn.taggings_table,
-              :tag_id,
-              algorithm: :concurrently unless index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
-    add_index ActsAsTaggableOn.taggings_table,
-              [:taggable_id, :taggable_type, :context],
+              %i[taggable_id taggable_type context],
               name: 'taggings_taggable_context_idx',
               algorithm: :concurrently
   end
